Changes to package.mk and branch-tool to support using the same makefiles for both CDH and CM.
branch-tool now takes --prefix=foo as an argument (rather than just
prefix=foo handled from argv), and has an optional --no-patch-count
flag. If the latter is provided, the version it spits out won't have a
patch count on it. This way, branch-tool can be used for CM as well as
for CDH. Additionally, cdh_get_version calls fixed to always use the
prefix, so that other branch-tool commands besides version will work.

In package.mk, this changes how branch-tool is called, of course. It
also adds a few more environment variables to be passed to
do-component-build and debuild, to support CM's build. But most
importantly, it's got some new logic to handle the three use cases we
have now for components:
1: The standard CDH components - pristine tarball, patch, build.
2: bigtop-utils - no tarball, no do-component-build.
3: CM - no tarball, but does still need do-component-build, no patching.

In a couple cases, we look to see if $($(PKG)_TARBALL_SRC) is set -
previously, that was used entirely to handle the bigtop-utils
case. Now it's only used for the cases where bigtop-utils and CM both
need to skip a step. For the cases where bigtop-utils should skip but
CM shouldn't, we leave $($(PKG)_TARBALL_DST) set for CM, and check if
that's unset - if so, as in the case for bigtop-utils, those steps
(i.e., do-component-build) are skipped. But for CM, they'll be
executed. There's also a check that ..._SRC is unset and ..._DST is
set - that's to catch the special case where CM has different behavior
from either bigtop-utils or the rest of CDH, e.g., copying the source
to the build area (for calling do-component-build) without any patch logic.

+++ Makefile
@@ -86,7 +86,7 @@ deb: install-pom $(TARGETS_DEB)
 relnotes: $(TARGETS_RELNOTES)
 
 install-pom:
-	mvn -N install $(DO_MAVEN_DEPLOY)
+	[ -f "$(BASE_DIR)/pom.xml" ] && mvn -N install $(DO_MAVEN_DEPLOY)
 
 .DEFAULT_GOAL:= help
 .PHONY: realclean clean distclean package-help help-header packages all world help tgz srpm sdeb install-pom relnotes
